requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Environment & Configuration
python-dotenv==1.0.0

# HTTP Requests & Web Scraping
requests==2.31.0
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11

# Google Gemini AI
google-generativeai==0.3.2

# Instagram Automation
instagrapi==2.1.2
Pillow==10.2.0

# Task Scheduling
schedule==1.2.1
python-crontab==3.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.39.2

# Utilities
pytz==2023.3
python-dateutil==2.8.2
validators==0.22.0

# Testing (Development)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx-mock==0.15.0

# Code Quality (Development)
black==23.12.1
flake8==7.0.0
mypy==1.8.0
